# apps/backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Credfolio API
  version: 1.0.0
  description: API for Smart Profile Generation and Credibility

paths:
  /profile:
    get:
      summary: Get current user's profile
      operationId: getProfile
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
              example:
                id: "123e4567-e89b-12d3-a456-426614174000"
                summary: "Experienced software engineer with expertise in full-stack development"
                workExperiences:
                  - id: "223e4567-e89b-12d3-a456-426614174001"
                    companyName: "Tech Corp"
                    role: "Senior Software Engineer"
                    startDate: "2020-01-01"
                    endDate: "2023-12-31"
                    description: "Led development of key features"
                    credibilityHighlights:
                      - quote: "Outstanding performance and leadership"
                        sentiment: "POSITIVE"
                skills: ["TypeScript", "React", "Node.js", "Go"]
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1001
                message: "Unauthorized"
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1201
                message: "Profile not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1501
                message: "Internal server error"
    put:
      summary: Update profile
      operationId: updateProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileInput'
            example:
              summary: "Updated professional summary"
      responses:
        '200':
          description: Updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1102
                message: "Invalid request body"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1001
                message: "Unauthorized"
        '404':
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1201
                message: "Profile not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1405
                message: "Failed to update profile"

  /reference-letters:
    post:
      summary: Upload a reference letter
      operationId: uploadReferenceLetter
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
            encoding:
              file:
                contentType: application/pdf
      responses:
        '201':
          description: Uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferenceLetter'
              example:
                id: "323e4567-e89b-12d3-a456-426614174002"
                fileName: "reference_letter.pdf"
                uploadDate: "2024-01-15T10:30:00Z"
                status: "PROCESSED"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1103
                message: "File is required"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1001
                message: "Unauthorized"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1401
                message: "Failed to process PDF"

  /profile/generate:
    post:
      summary: Generate profile from uploaded letters
      operationId: generateProfile
      responses:
        '200':
          description: Generation started/completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1301
                message: "No reference letters found"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1001
                message: "Unauthorized"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1403
                message: "Failed to generate profile"

  /profile/tailor:
    post:
      summary: Create a tailored profile for a job description
      operationId: tailorProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                jobDescription:
                  type: string
            example:
              jobDescription: "We are looking for a Senior Software Engineer with experience in React, TypeScript, and Node.js. The ideal candidate should have 5+ years of experience in full-stack development."
      responses:
        '200':
          description: Tailored profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobMatch'
              example:
                id: "423e4567-e89b-12d3-a456-426614174003"
                matchScore: 0.85
                tailoredSummary: "Highly qualified candidate with strong alignment to job requirements"
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1302
                message: "Job description is required"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1001
                message: "Unauthorized"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1404
                message: "Failed to tailor profile"

  /profile/{profileId}/cv:
    get:
      summary: Download CV as PDF
      operationId: downloadCV
      parameters:
        - name: profileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        - name: jobMatchId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Optional job match ID for tailored CV
          example: "423e4567-e89b-12d3-a456-426614174003"
      responses:
        '200':
          description: PDF File
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1106
                message: "Invalid jobMatchId"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1001
                message: "Unauthorized"
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1002
                message: "Forbidden"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1201
                message: "Profile not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_id: 1402
                message: "Failed to generate CV"

components:
  schemas:
    Profile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        summary:
          type: string
        workExperiences:
          type: array
          items:
            $ref: '#/components/schemas/WorkExperience'
        skills:
          type: array
          items:
            type: string

    ProfileInput:
      type: object
      properties:
        summary:
          type: string
        workExperiences:
          type: array
          items:
            $ref: '#/components/schemas/WorkExperienceInput'
        skills:
          type: array
          items:
            type: string

    WorkExperience:
      type: object
      properties:
        id:
          type: string
          format: uuid
        companyName:
          type: string
        role:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        description:
          type: string
        credibilityHighlights:
          type: array
          items:
            $ref: '#/components/schemas/CredibilityHighlight'

    WorkExperienceInput:
      type: object
      properties:
        companyName:
          type: string
        role:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        description:
          type: string

    CredibilityHighlight:
      type: object
      properties:
        quote:
          type: string
        sentiment:
          type: string
          enum: [POSITIVE, NEUTRAL]

    ReferenceLetter:
      type: object
      properties:
        id:
          type: string
          format: uuid
        fileName:
          type: string
        uploadDate:
          type: string
          format: date-time
        status:
          type: string

    JobMatch:
      type: object
      properties:
        id:
          type: string
          format: uuid
        matchScore:
          type: number
          format: float
          minimum: 0
          maximum: 1
        tailoredSummary:
          type: string

    ErrorResponse:
      type: object
      required:
        - error_id
        - message
      properties:
        error_id:
          type: integer
          description: Numeric error identifier for programmatic error handling
          example: 1301
        message:
          type: string
          description: Human-readable error message
          example: "No reference letters found"